chore: merge pydantic_ai file to simplify.
